possible improvements
Changed the 28% extra fee in Front Running to 30%. As it was somehow not working while creating the demonstration video.

Also note that higher prices cause the overall block fee to rise. Thus, the initial transaction which is to be front run, that itself doesn't get deployed.
While it is still a type of front running (suppression). This was not what we were looking for.